refactor: update logo imports to new svg directory structure
- Changed logo import paths in ForgotPassword, Login, Signup, and VerifyEmail components to reflect the new directory structure for SVG assets.
- Ensured consistency in asset management across the unauthenticated pages.
